refactor(velociraptor): move the Zone.Identifier check into its own module
Merging master added #692's work to the same file, and the two together pushed
velociraptorImport.ts past the size the ledger freezes it at. The zone rules are
self-contained, so they leave rather than the ledger moving.

Also closes a test block that the merge cut in half. Both branches appended to
the end of velociraptorImport.test.ts, and the conflict boundary landed inside
the last case; the file parsed only after the block was closed.
